requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

langchain==0.0.352
pandas==2.1.4
passlib==1.7.4
Pillow==9.5.0
plotly==5.18.0
pymongo==4.6.1
requests==2.28.2
streamlit==1.25.0
